Set up Actix Web server for PDF creation
Receives HTTP requests with a JSON body containing data for the `Parameters` struct.

The data is deserialized and an HTML document is rendered with it.

The HTML is used by `wkhtmltopdf` to create the PDF document, which is saved locally.

A severe limitation of `wkhtmltopdf` is that it only allows one initialization of `PdfApplication`
per process, and all PDF generation must happen from the thread that initialized it.

This means that I had to set the number of workers for
the `HttpServer` to `1`, so that only one `actix_web::App` instance was created in just one thread.